+++ library/entropy_poll.h
